Headless doesn't need pyqt at all now & optimizations
